refactor: /types/user.ts 분리
