+++ src/types/common.ts
@@ -1,0 +1,5 @@
+export interface ApiResponse<T> {
+  data: T;
+  message: string;
+  success: boolean;
+}
